Separate de-id Pipiline Options in a different file
